+++ project/pokemonsApp/src/app/pokemon/list-pokemon/list-pokemon.component.ts
@@ -1,16 +1,23 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Pokemon } from '../interfaces/pokemon.js';
-import { POKEMONS } from '../mock-pokemons.js';
+import { PokemonService } from '../pokemon.service.js';
 
 @Component({
   selector: 'app-list-pokemon',
   templateUrl: './list-pokemon.component.html',
 })
-export class ListPokemonComponent {
-  pokemonList: Pokemon[] = POKEMONS;
+export class ListPokemonComponent implements OnInit {
+  pokemonList: Pokemon[];
 
-  constructor(private router: Router) {}
+  constructor(private router: Router, private pokemonService: PokemonService) {}
+
+  ngOnInit(): void {
+    // this.pokemonList = this.pokemonService.getPokemonList();
+    this.pokemonService
+      .getPokemonList()
+      .subscribe((pokemons) => (this.pokemonList = pokemons));
+  }
 
   goToPokemonSelected(pokemon: Pokemon) {
     this.router.navigate(['pokemon', pokemon.id]);